+++ tests/Jroc.Test262.Tests/built-ins/String/prototype/toUpperCase/ExecutionTests.UnicodeCasing.cs
@@ -1,0 +1,7 @@
+namespace Jroc.Test262.Tests.built_ins.String.prototype.toUpperCase;
+
+public partial class ExecutionTests
+{
+    [Fact(DisplayName = "special_casing.js")]
+    public Task special_casing() => ExecutionTestFromFile("special_casing");
+}